Crash on exit when overriding signals handlers in states.
It appears that it is still possible to have an exit crash even after
QTBUG-21617. See

    tests/auto/declarative/qdeclarativestates/tst_qdeclarativestates.cpp

for an analysis of the issue.

The short term solution: use reference counted pointers for the signal
handler expressions. Considering all the cases we want to preserve an
expression:

  - State is destroyed but the item is alive an is in this state
  - State is destroyed but the item has a reverse action in its reverse
    list uses this expression (not sure it is actually usable)
  - The original item signal handler expression should be preserved
    regardless of the state we are in and regardless of what do we do
    with the states (add new, delete, change current state)

It seems like reference counting is the only manageable solution unless
the overall model is modified a little bit.

In the long run I would advice a more general refactoring with a strict
rules of ownership for the expressions, state groups and states. I
would say that the reverse list should be moved into the state group and
states should own their expressions. And we do not allow destruction of
the current state or any of the states current state extends.

When a state is destroyed we would delete all the expressions that
belong to that state. The reverse list should contain only references
to the expressions that were the initial values of the item signal
handlers. Thus unless we are trying to remove current state (or a state
that the current state extends) we should be fine from the semantic
standpoint.