Disabled buttons should not consume mouse wheel events
As buttons normally ignore all mouse wheel events it is wired if it
starts consuming then when disabled.

It seems that when another bug was fixed the mouse wheel event was
included in the list by accident.  All the other events that a button
normally consume should be consumed the same way when it is disabled.