Fixed localization of AM/PM strings in QDateTime::getFmtString()
Prior to this change, calls to QDateTime::getFmtString() would
replace the AP token in a time format with hard-coded "am" or "pm"
values, regardless of the current locale. This was evident in calls
to QDateTime::timeString().

This fix replaces the hard-coded strings with calls to
QLocale::amText() and QLocale::pmText(), which return the correct
localized strings for "am" and "pm", respectively.

Reviewer: sraub