Remove call to ClipCursor in enterModal_sys()
This call has adverse affects on unrelated fullscreen applications,
in that it will allow the mouse cursor to exit a fullscreen
application's window boundaries once invoked where prior to the
mouse was constrained to that application's window dimensions.

This is implementation-specific code that should be invoked at the
application level (outside of Qt) if the use case demands it.